+++ plugins/techdocs-node/src/stages/publish/googleStorage.ts
@@ -26,10 +26,10 @@ import express from 'express';
 import JSON5 from 'json5';
 import path from 'path';
 import { Readable } from 'stream';
-import { Logger } from 'winston';
 import {
   getFileTreeRecursively,
   getHeadersForFileExtension,
+  isValidContentPath,
   lowerCaseEntityTriplet,
   lowerCaseEntityTripletInStoragePath,
   bulkStorageOperation,
@@ -45,19 +45,20 @@ import {
   ReadinessResponse,
   TechDocsMetadata,
 } from './types';
+import { LoggerService } from '@backstage/backend-plugin-api';
 
 export class GoogleGCSPublish implements PublisherBase {
   private readonly storageClient: Storage;
   private readonly bucketName: string;
   private readonly legacyPathCasing: boolean;
-  private readonly logger: Logger;
+  private readonly logger: LoggerService;
   private readonly bucketRootPath: string;
 
   constructor(options: {
     storageClient: Storage;
     bucketName: string;
     legacyPathCasing: boolean;
-    logger: Logger;
+    logger: LoggerService;
     bucketRootPath: string;
   }) {
     this.storageClient = options.storageClient;
@@ -67,7 +68,7 @@ export class GoogleGCSPublish implements PublisherBase {
     this.bucketRootPath = options.bucketRootPath;
   }
 
-  static fromConfig(config: Config, logger: Logger): PublisherBase {
+  static fromConfig(config: Config, logger: LoggerService): PublisherBase {
     let bucketName = '';
     try {
       bucketName = config.getString('techdocs.publisher.googleGcs.bucketName');
@@ -262,6 +263,12 @@ export class GoogleGCSPublish implements PublisherBase {
         : lowerCaseEntityTriplet(entityTriplet);
 
       const entityRootDir = path.posix.join(this.bucketRootPath, entityDir);
+      if (!isValidContentPath(this.bucketRootPath, entityRootDir)) {
+        this.logger.error(
+          `Invalid content path found while fetching TechDocs metadata: ${entityRootDir}`,
+        );
+        reject(new Error(`Metadata Not Found`));
+      }
 
       const fileStreamChunks: Array<any> = [];
       this.storageClient
@@ -297,6 +304,13 @@ export class GoogleGCSPublish implements PublisherBase {
 
       // Prepend the root path to the relative file path
       const filePath = path.posix.join(this.bucketRootPath, filePathNoRoot);
+      if (!isValidContentPath(this.bucketRootPath, filePath)) {
+        this.logger.error(
+          `Attempted to fetch TechDocs content for a file outside of the bucket root: ${filePathNoRoot}`,
+        );
+        res.status(404).send('File Not Found');
+        return;
+      }
 
       // Files with different extensions (CSS, HTML) need to be served with different headers
       const fileExtension = path.extname(filePath);
@@ -337,6 +351,12 @@ export class GoogleGCSPublish implements PublisherBase {
         : lowerCaseEntityTriplet(entityTriplet);
 
       const entityRootDir = path.posix.join(this.bucketRootPath, entityDir);
+      if (!isValidContentPath(this.bucketRootPath, entityRootDir)) {
+        this.logger.error(
+          `Invalid content path found while checking if docs have been generated: ${entityRootDir}`,
+        );
+        resolve(false);
+      }
 
       this.storageClient
         .bucket(this.bucketName)