+++ plugins/techdocs-node/src/stages/publish/index.ts
@@ -18,10 +18,10 @@ export type {
   PublisherBase,
   PublisherType,
   PublisherFactory,
+  PublisherBuilder,
   PublishRequest,
   PublishResponse,
   MigrateRequest,
   ReadinessResponse,
   TechDocsMetadata,
-  PublisherBuilder,
 } from './types';